fix overlapping borders
when a floating or transient client exists,
then, selecting normal clients wouldn't raise them,
so, its active border was hidden, most times.
now every window is lowered, except the active one,
so the its borders are always on top.

this also fixes flickering of floating/transient
clients that were ovelapping.

however it introduces a weird behavior
when the mode is monocle, as in that case
when the floating client is selected,
the background/fullscreen window is always
the first opened window (or head).