fix no current client when removing prevfocus
for example we open an application with a splash screen.
the splash screen is the current window until the app
is ready, opened, and mapped.
once it's ready, the splash screen, which is now prevfocus
is destroyed, and thus prevfocus is null.
closing the app, called update_current(prevfocus) which
resulted in no client being current and highlighted.

also the check to remove _NET_ACTIVE_WINDOW is done
with head. if there is no head, there is no client.