set _NET_ACTIVE_WINDOW when requested
tools like `xdotool` can be used to select/focus
a window, using the _NET_ACTIVE_WINDOW property.
however those tools don't know or don't care about
the state of the window, ie the window is unmapped.
we workaround this by checking if the client that
requested that property is visible in the current
desktop. If not we discard the client message.