bigfix focus behavior on FLOAT mode
clients spawned when mode was FLOAT would not be
marked as floating clients, and thus floating clients
would appear on top of them.
now clients spawned when mode is FLOAT, are marked as
floating clients.

additionally, there is no 'reset floating clients'
capability, when mode is FLOAT. this was achieved by
retiling the windows (reselecting the current mode).
it would mark floating clients as non-floating and thus
tile/reset them to the layout.
this must not happen when the mode is FLOAT, as if
a client on FLOAT mode is marked non-floating and
moved to another desktop, it will be tiled, and this
is inconsistent behavior (unexpected result).

clients on FLOAT mode are floating clients always, and
switching to FLOAT mode should mark all clients as such.