fix CLICK_TO_FOCUS behavior
when CLICK_TO_FOCUS was set, a Button1 click on a client
should make that client the current one (highlight and
input focus). however this was inconsistent, as if a
modifier key was pressed at that time, like CapsLock or
NumLock, the click would not be grabbed.
this commit fixes this behavior and CLICK_TO_FOCUS should
work regardless of NumLock or CapsLock or both modifiers
beeing pressed at the time of the click event.